+++ src/components/numberbox/Numberbox.tsx
@@ -1,5 +1,5 @@
-import { useState } from "react";
 import "./Numberbox.css";
+import NumberInput from "./NumberInput";
 
 interface Props {
   value: number;
@@ -11,23 +11,13 @@ interface Props {
 export const Numberbox = (props: Props) => {
   const { value, setValue, label, hint } = props;
 
-  const [internalValue, setInternalValue] = useState(value);
-
   return (
     <label className="textbox" title={hint}>
       {label}
-      <input
-        type="number"
+      <NumberInput
         className="textbox-input"
-        step="0.1"
-        value={internalValue}
-        onChange={(e) => setInternalValue(Number(e.target.value))}
-        onBlur={() => setValue(internalValue)}
-        onKeyDown={(e) => {
-          if (e.key === "Enter") {
-            (e.target as HTMLInputElement).blur();
-          }
-        }}
+        value={value}
+        setValue={setValue}
       />
     </label>
   );